(doc temp) Temporarily removing all documentation.
Quite a lot of the documentation (such as it was) was getting very out-of-sync with the Nuclear design itself, to
say nothing of the actual codebase. It all needs to be rewritten, but I can’t stop to do that right now; so I’m
removing it outright pending a rewrite.

A lot of this will go back in verbatim, or with minor changes; it just needs ratification as “still applicable.”