(- meta fix) Referring people to #elliottcable instead of #Paws.Nucleus
This isn’t the *official* channel, but it’s the preferred one for the moment. Until Freenode gives us ownership
of #Paws, and there’s enough of a language to worry about having its own separate community from my own, it’s
easier to concentrate everything in a single channel.